requirements.txt
# Document Processing
pypdf2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Text Processing
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2020.1.16

# Utilities
python-magic>=0.4.27
chardet>=5.2.0

# Embeddings & Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# FastAPI & Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Retrieval
rank-bm25>=0.2.2

# Reranking
FlagEmbedding>=1.2.0

# LLM
openai>=1.0.0

# Environment Variables
python-dotenv>=1.0.0

